--- src/libraries/Microsoft.Win32.Registry/ref/Microsoft.Win32.Registry.net461.cs
@@ -1,21 +1,0 @@
-// Licensed to the .NET Foundation under one or more agreements.
-// The .NET Foundation licenses this file to you under the MIT license.
-// ------------------------------------------------------------------------------
-// Changes to this file must follow the https://aka.ms/api-review process.
-// ------------------------------------------------------------------------------
-
-using System.Runtime.CompilerServices;
-
-[assembly: TypeForwardedTo(typeof(System.Security.AccessControl.RegistryAccessRule))]
-[assembly: TypeForwardedTo(typeof(System.Security.AccessControl.RegistryAuditRule))]
-[assembly: TypeForwardedTo(typeof(System.Security.AccessControl.RegistryRights))]
-[assembly: TypeForwardedTo(typeof(System.Security.AccessControl.RegistrySecurity))]
-[assembly: TypeForwardedTo(typeof(Microsoft.Win32.Registry))]
-[assembly: TypeForwardedTo(typeof(Microsoft.Win32.RegistryHive))]
-[assembly: TypeForwardedTo(typeof(Microsoft.Win32.RegistryKey))]
-[assembly: TypeForwardedTo(typeof(Microsoft.Win32.RegistryKeyPermissionCheck))]
-[assembly: TypeForwardedTo(typeof(Microsoft.Win32.RegistryOptions))]
-[assembly: TypeForwardedTo(typeof(Microsoft.Win32.RegistryValueKind))]
-[assembly: TypeForwardedTo(typeof(Microsoft.Win32.RegistryValueOptions))]
-[assembly: TypeForwardedTo(typeof(Microsoft.Win32.RegistryView))]
-[assembly: TypeForwardedTo(typeof(Microsoft.Win32.SafeHandles.SafeRegistryHandle))]
