--- src/libraries/System.IO.FileSystem.AccessControl/ref/System.IO.FileSystem.AccessControl.net461.cs
@@ -1,15 +1,0 @@
-// Licensed to the .NET Foundation under one or more agreements.
-// The .NET Foundation licenses this file to you under the MIT license.
-// ------------------------------------------------------------------------------
-// Changes to this file must follow the https://aka.ms/api-review process.
-// ------------------------------------------------------------------------------
-
-using System.Runtime.CompilerServices;
-
-[assembly: TypeForwardedTo(typeof(System.Security.AccessControl.DirectoryObjectSecurity))]
-[assembly: TypeForwardedTo(typeof(System.Security.AccessControl.DirectorySecurity))]
-[assembly: TypeForwardedTo(typeof(System.Security.AccessControl.FileSecurity))]
-[assembly: TypeForwardedTo(typeof(System.Security.AccessControl.FileSystemAccessRule))]
-[assembly: TypeForwardedTo(typeof(System.Security.AccessControl.FileSystemAuditRule))]
-[assembly: TypeForwardedTo(typeof(System.Security.AccessControl.FileSystemRights))]
-[assembly: TypeForwardedTo(typeof(System.Security.AccessControl.FileSystemSecurity))]
