+++ src/libraries/System.IO.Pipes.AccessControl/Directory.Build.props
@@ -3,12 +3,5 @@
   <PropertyGroup>
     <StrongNameKeyId>Microsoft</StrongNameKeyId>
     <SupportedOSPlatforms>windows</SupportedOSPlatforms>
-    <PackageDescription>Provides types for managing access and audit control lists for pipes.
-
-Commonly Used Types:
-System.IO.Pipes.PipeSecurity
-System.IO.Pipes.PipeAccessRule
-System.IO.Pipes.PipeAuditRule
-System.IO.Pipes.PipeAccessRights</PackageDescription>
   </PropertyGroup>
 </Project>
