--- src/libraries/System.IO.Pipes.AccessControl/ref/System.IO.Pipes.AccessControl.net461.cs
@@ -1,12 +1,0 @@
-// Licensed to the .NET Foundation under one or more agreements.
-// The .NET Foundation licenses this file to you under the MIT license.
-// ------------------------------------------------------------------------------
-// Changes to this file must follow the https://aka.ms/api-review process.
-// ------------------------------------------------------------------------------
-
-using System.Runtime.CompilerServices;
-
-[assembly: TypeForwardedTo(typeof(System.IO.Pipes.PipeAccessRights))]
-[assembly: TypeForwardedTo(typeof(System.IO.Pipes.PipeAccessRule))]
-[assembly: TypeForwardedTo(typeof(System.IO.Pipes.PipeAuditRule))]
-[assembly: TypeForwardedTo(typeof(System.IO.Pipes.PipeSecurity))]
