--- src/libraries/System.Security.AccessControl/NuGet.config
@@ -1,12 +1,0 @@
-<configuration>
-  <!-- This disables restore in VS for projects that target platform specific target frameworks older than net5.0. -->
-  <packageRestore>
-    <!-- The 'automatic' key is set to True when the "Automatically check for missing packages during
-         build in Visual Studio" checkbox is set. Clearing the box sets this to False and disables
-         automatic restore. -->
-    <add key="automatic" value="False" />
-    <!-- The 'enabled' key is True when the "Allow NuGet to download missing packages" checkbox is set.
-         Clearing the box sets this to False, disabling command-line, automatic, and MSBuild-integrated restore. -->
-    <add key="enabled" value="False" />
-  </packageRestore>
-</configuration>
