--- src/libraries/System.Security.Principal.Windows/ref/System.Security.Principal.Windows.net461.cs
@@ -1,20 +1,0 @@
-// Licensed to the .NET Foundation under one or more agreements.
-// The .NET Foundation licenses this file to you under the MIT license.
-// ------------------------------------------------------------------------------
-// Changes to this file must follow the https://aka.ms/api-review process.
-// ------------------------------------------------------------------------------
-
-using System.Runtime.CompilerServices;
-
-[assembly: TypeForwardedTo(typeof(System.Security.Principal.IdentityNotMappedException))]
-[assembly: TypeForwardedTo(typeof(System.Security.Principal.IdentityReference))]
-[assembly: TypeForwardedTo(typeof(System.Security.Principal.IdentityReferenceCollection))]
-[assembly: TypeForwardedTo(typeof(System.Security.Principal.NTAccount))]
-[assembly: TypeForwardedTo(typeof(System.Security.Principal.SecurityIdentifier))]
-[assembly: TypeForwardedTo(typeof(System.Security.Principal.TokenAccessLevels))]
-[assembly: TypeForwardedTo(typeof(System.Security.Principal.WellKnownSidType))]
-[assembly: TypeForwardedTo(typeof(System.Security.Principal.WindowsAccountType))]
-[assembly: TypeForwardedTo(typeof(System.Security.Principal.WindowsBuiltInRole))]
-[assembly: TypeForwardedTo(typeof(System.Security.Principal.WindowsIdentity))]
-[assembly: TypeForwardedTo(typeof(System.Security.Principal.WindowsPrincipal))]
-[assembly: TypeForwardedTo(typeof(Microsoft.Win32.SafeHandles.SafeAccessTokenHandle))]
