+++ src/libraries/System.Security.Principal.Windows/src/System/Security/Principal/SID.cs
@@ -314,11 +314,7 @@ namespace System.Security.Principal
 
         [MemberNotNull(nameof(_binaryForm))]
         [MemberNotNull(nameof(_subAuthorities))]
-#if NETSTANDARD2_0
-        private void CreateFromParts(IdentifierAuthority identifierAuthority, int[] subAuthorities)
-#else
         private void CreateFromParts(IdentifierAuthority identifierAuthority, ReadOnlySpan<int> subAuthorities)
-#endif
         {
             //
             // Check the number of subauthorities passed in
@@ -350,11 +346,7 @@ namespace System.Security.Principal
             //
 
             _identifierAuthority = identifierAuthority;
-#if NETSTANDARD2_0
-            _subAuthorities = (int[])subAuthorities.Clone();
-#else
             _subAuthorities = subAuthorities.ToArray();
-#endif
 
             //
             // Compute and store the binary form
@@ -462,12 +454,7 @@ namespace System.Security.Principal
                 throw new ArgumentException(SR.ArgumentOutOfRange_ArrayTooSmall, nameof(binaryForm));
             }
 
-#if NETSTANDARD2_0
-            int[] subAuthorities = new int[subAuthoritiesLength];
-#else
             Span<int> subAuthorities = stackalloc int[MaxSubAuthorities];
-#endif
-
             IdentifierAuthority authority = (IdentifierAuthority)(
                 (((long)binaryForm[offset + 2]) << 40) +
                 (((long)binaryForm[offset + 3]) << 32) +
@@ -494,11 +481,7 @@ namespace System.Security.Principal
 
             CreateFromParts(
                 authority,
-#if NETSTANDARD2_0
-                subAuthorities
-#else
                 subAuthorities.Slice(0, subAuthoritiesLength)
-#endif
             );
 
             return;
@@ -657,11 +640,7 @@ namespace System.Security.Principal
             CreateFromBinaryForm(resultSid!, 0);
         }
 
-#if NETSTANDARD2_0
-        internal SecurityIdentifier(IdentifierAuthority identifierAuthority, int[] subAuthorities)
-#else
         internal SecurityIdentifier(IdentifierAuthority identifierAuthority, ReadOnlySpan<int> subAuthorities)
-#endif
         {
             CreateFromParts(identifierAuthority, subAuthorities);
         }
@@ -748,15 +727,6 @@ namespace System.Security.Principal
                 // otherwise you would see this: "S-1-NTAuthority-32-544"
                 //
 
-#if NETSTANDARD2_0
-                StringBuilder result = new StringBuilder();
-                result.Append("S-1-").Append((ulong)_identifierAuthority);
-                for (int i = 0; i < SubAuthorityCount; i++)
-                {
-                    result.Append('-').Append((uint)(_subAuthorities[i]));
-                }
-                _sddlForm = result.ToString();
-#else
                 // length of buffer calculation
                 // prefix = "S-1-".Length: 4;
                 // authority: ulong.MaxValue.ToString("D") : 20;
@@ -779,7 +749,6 @@ namespace System.Security.Principal
                     length += written;
                 }
                 _sddlForm = result.Slice(0, length).ToString();
-#endif
             }
 
             return _sddlForm;
